Add liquidity calculation scaffolding for StablePhantom and MetaStable
