+++ balancer-js/src/modules/pools/pool-types/concerns/meta-stable/liquidity.concern.ts
@@ -2,7 +2,7 @@ import { LiquidityConcern } from '../types';
 import { TokenBalance } from '@/types';
 import { BigNumber, parseFixed, formatFixed } from '@ethersproject/bignumber';
 
-export class StablePoolLiquidity implements LiquidityConcern {
+export class MetaStablePoolLiquidity implements LiquidityConcern {
     calcTotal(tokenBalances: TokenBalance[]): string {
         // TODO implementation
         console.log(tokenBalances);
