+++ balancer-js/src/modules/pools/pool-types/concerns/stable-phantom/liquidity.concern.ts
@@ -2,7 +2,7 @@ import { LiquidityConcern } from '../types';
 import { TokenBalance } from '@/types';
 import { BigNumber, parseFixed, formatFixed } from '@ethersproject/bignumber';
 
-export class StablePoolLiquidity implements LiquidityConcern {
+export class StablePhantomPoolLiquidity implements LiquidityConcern {
     calcTotal(tokenBalances: TokenBalance[]): string {
         // TODO implementation
         console.log(tokenBalances);
