+++ balancer-js/src/modules/pools/pool-types/concerns/weighted/liquidity.concern.ts
@@ -1,21 +1,15 @@
 import { LiquidityConcern } from '../types';
-import { BigNumber, parseFixed, formatFixed, BigNumberish } from '@ethersproject/bignumber';
-
-export interface TokenBalance {
-    balance: string;
-    decimals: number;
-    priceRate: string;
-    price: string;
-}
+import { TokenBalance } from '@/types';
+import { BigNumber, parseFixed, formatFixed } from '@ethersproject/bignumber';
 
 export class WeightedPoolLiquidity implements LiquidityConcern {
     calcTotal(tokenBalances: TokenBalance[]): string {
         let sumValue = BigNumber.from(0);
 
         for (let i = 0; i < tokenBalances.length; i++) {
-            const token = tokenBalances[i];
-            const price = parseFixed(token.price, 18);
-            const balance = parseFixed(token.balance, 18);
+            const tokenBalance = tokenBalances[i];
+            const price = parseFixed(tokenBalance.token.price, 18);
+            const balance = parseFixed(tokenBalance.balance, 18);
 
             const value = balance.mul(price);
             sumValue = sumValue.add(value);
