+++ balancer-js/src/modules/pools/pools.module.ts
@@ -1,4 +1,5 @@
 import { BalancerSdkConfig } from '@/types';
+import { PoolType } from './types';
 import { Stable } from './pool-types/stable.module';
 import { Weighted } from './pool-types/weighted.module';
 import { MetaStable } from './pool-types/metaStable.module';
